+ Isolated Storage(partial) + touch-safe button
We are VERY close on having a fully workout Isolated Storage database!
Also, End workout button now has an ethical, touch-safe mechanism in
order to prevent accidental workout session termination.

There is NOW an isolated storage code implementation.

The 'DataSaver' class takes care of loading and saving the data to the
isolated storage.

The loading and saving implementation works HOWEVER. For some reason,
when you close the application, the saved database still does not appear
to 'stay around' in the isolated storage.

Here's my list of possible reasons:
- The database gets wiped out after closing the application.
- Workouts are added in the wrong database( is possibly saved in a
hardcoded database that I haven't noticed perhaps)
- Database is not loaded properly in listbox on the
'PreviousWorkoutListPage'
- The database that is used to update the listbox with is not the
correct one.(Possibly a hard-coded database is used instead, which
ofcourse, always gets wiped out after closing the application)

If we can fix this, we are soooo done with this application!! :)

Well aside of the fact that we also have to display a freaking graph
displaying workout speed and distance values.